# Use a call-graph in the inliner