--- a/sway-ir/src/analysis.rs
+++ b/sway-ir/src/analysis.rs
@@ -1,2 +1,4 @@
+pub mod call_graph;
+pub use call_graph::*;
 pub mod dominator;
 pub use dominator::*;