@vaivaswatha 2022-10-28
> I guess the bytecode changes are just in the ordering of the function bodies in the bytecode, then?
> 
> Awesome PR, love to see the performance boost!

Suppose A calls B and B calls C. We have a limit on how big a function we can inline.
So if I inline C into B first, B may now be too big to be inlined into A. But if i inline B into A first, I can still inline C into A after that.

What this PR does is to introduce a "callee first" order of traversing the program for inlining. This ensures that you don't have to inline the same call site multiple times. But because the order of visiting call sites  has changed (compared to before this PR), there may be some differences in the final inlining decisions.